+++ lib/ui/border.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:ones/core/copywith.dart';
 
 extension BorderX<T extends Widget> on T {
   ///add border
-  Container border({Color color = const Color(0xff95d9f1), double width = 2, double radius = 0}) {
+  Container border({Color color = const Color(0xff95d9f1), double width = 1, double radius = 0}) {
     return Container(
         decoration: BoxDecoration(
           border: Border.all(color: color, width: width),
@@ -14,17 +15,38 @@ extension BorderX<T extends Widget> on T {
   }
 
   ///add Circle border
-  Container circleBorder({Color color = const Color(0xff95d9f1), double width = 2}) {
+  Container circleBorder({Color color = const Color(0xff95d9f1), double width = 1}) {
+    return border(color: color, width: width, radius: 1000);
+  }
+}
 
+///Container border
+extension BorderContainerX on Container {
 
-
-    return Container(
-        decoration: BoxDecoration(
-          border: Border.all(color: color, width: width),
-          borderRadius: const BorderRadius.all(
-            Radius.circular(1000),
-          ),
+  ///container border
+  Container border({Color color = const Color(0xff95d9f1), double width = 1, double radius = 0}) {
+    late Decoration mergeDecoration;
+    if (decoration is BoxDecoration) {
+      mergeDecoration = (decoration as BoxDecoration).copyWith(
+        border: Border.all(color: color, width: width),
+        borderRadius: BorderRadius.all(
+          Radius.circular(radius),
         ),
-        child: this);
+      );
+    } else {
+      mergeDecoration = BoxDecoration(
+        border: Border.all(color: color, width: width),
+        borderRadius: BorderRadius.all(
+          Radius.circular(radius),
+        ),
+      );
+    }
+
+    return copyWith(decoration: mergeDecoration);
+  }
+
+  ///add Circle border
+  Container circleBorder({Color color = const Color(0xff95d9f1), double width = 1}) {
+    return border(color: color, width: width, radius: 1000);
   }
 }